refactor: extract module inlining logic into dedicated inliner module
Refactored the bundler.rs file to improve code organization by extracting all module inlining functionality into a new dedicated inliner.rs module.

Changes:
- Created new inliner.rs module (748 lines) containing all inlining logic
- Moved inline_module, inline_class, inline_assignment, and inline_ann_assignment functions
- Extracted main inlining loop into new inline_all_modules function
- Updated bundle_modules to use the new inliner module
- Made necessary helper methods public for cross-module access
- Reduced bundler.rs from 7,671 to 6,950 lines (721 lines reduction)

All tests pass and clippy reports no warnings. This refactoring improves maintainability without changing any functionality.
