+++ crates/cribo/src/code_generator/mod.rs
@@ -13,6 +13,7 @@ pub mod expression_handlers;
 pub mod globals;
 pub mod import_deduplicator;
 pub mod import_transformer;
+pub mod inliner;
 pub mod module_registry;
 pub mod module_transformer;
 pub mod namespace_manager;
